feat
Remaining interface tests are added, a new listener is added to assemble evidence in doc and xls files. Operator testing begins to be implemented

+++ src/main/java/dagserver/utils/ReportListener.java
@@ -4,34 +4,34 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
+import java.util.Collections;
 import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.Properties;
+
 import org.apache.poi.ss.usermodel.Row;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
 import org.apache.poi.util.Units;
 import org.apache.poi.xwpf.usermodel.ParagraphAlignment;
-import org.apache.poi.xwpf.usermodel.UnderlinePatterns;
 import org.apache.poi.xwpf.usermodel.XWPFDocument;
 import org.apache.poi.xwpf.usermodel.XWPFParagraph;
 import org.apache.poi.xwpf.usermodel.XWPFRun;
 import org.apache.poi.xwpf.usermodel.XWPFTable;
 import org.apache.poi.xwpf.usermodel.XWPFTableRow;
+import org.testng.IReporter;
 import org.testng.ISuite;
-import org.testng.ISuiteListener;
+import org.testng.xml.XmlSuite;
 
-public class MyListener implements ISuiteListener {
+public class ReportListener implements IReporter {
 
-	
 	protected String reportpath;
 	protected String evidencepath;
 	protected SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMddhhmmss");
@@ -40,12 +40,13 @@ public class MyListener implements ISuiteListener {
 	protected XWPFDocument document;
 	protected Sheet sheet;
 	
-	@Override
-	public void onStart(ISuite suite2) {
+	
+	public ReportListener(){
 		try {
-			
-			this.reportpath = "C:\\tmp\\test\\report_"+sdf.format(new Date())+".xls";
-			this.evidencepath = "C:\\tmp\\test\\evidence_"+sdf.format(new Date())+".doc";
+			Properties props = this.getProperties();
+			String reportpath = props.getProperty("params.reportpath");
+			this.reportpath = reportpath + "report_"+sdf.format(new Date())+".xls";
+			this.evidencepath = reportpath + "evidence_"+sdf.format(new Date())+".doc";
 			FileInputStream inputStream = new FileInputStream(new File("./src/main/resources/template_evidence_report.xlsx"));
 			workbook = WorkbookFactory.create(inputStream);
 			String titleStr = "Testing Cycle " + sdfst.format(new Date());
@@ -63,23 +64,44 @@ public class MyListener implements ISuiteListener {
 			workbook.setSheetName(0, "Testing Cycle " + sdfst.format(new Date()));
 			sheet = workbook.getSheetAt(0);
 		} catch (Exception e) {
-			e.printStackTrace();
+			// TODO: handle exception
 		}
 	}
 	
+	
+	private Properties getProperties() throws IOException {
+		Properties properties = new Properties();
+        String propFileName = "application.properties";
+
+        InputStream inputStream = this.getClass().getClassLoader().getResourceAsStream(propFileName);
+
+        if (inputStream != null) {
+            properties.load(inputStream);
+        } else {
+            throw new RuntimeException("Archivo de propiedades '" + propFileName + "' no se encuentra en el classpath.");
+        }
+        return properties;
+	}
+	
 	@Override
-	public void onFinish(ISuite suite1) {
+	public void generateReport(List<XmlSuite> xmlSuites, List<ISuite> suites, String outputDirectory) {
 		try {
-			List<Map<String,String>> list = (List<Map<String, String>>) suite1.getAttribute("dataOutput");
-			for (Iterator<Map<String, String>> iterator = list.iterator(); iterator.hasNext();) {
-				Map<String, String> map = iterator.next();
-				String name = map.get("name"); 
-				String classname = map.get("classname"); 
-				String status = map.get("status");
-				String screens = map.get("screenshoot");
-				this.writeEvidenceXls(name, classname, status);
-				this.writeEvidenceDoc(name, classname, status, screens);
-				new File(map.get("screenshoot")).delete();
+			Collections.reverse(suites);
+			for (Iterator<ISuite> iterator = suites.iterator(); iterator.hasNext();) {
+				ISuite iSuite = iterator.next();
+				List<Map<String,String>> list = (List<Map<String, String>>) iSuite.getAttribute("dataOutput");
+				if(list != null) {
+					for (Iterator<Map<String, String>> iterator2 = list.iterator(); iterator2.hasNext();) {
+						Map<String, String> map = iterator2.next();
+						String name = map.get("name"); 
+						String classname = map.get("classname"); 
+						String status = map.get("status");
+						String screens = map.get("screenshoot");
+						this.writeEvidenceXls(name, classname, status);
+						this.writeEvidenceDoc(name, classname, status, screens);
+						new File(map.get("screenshoot")).delete();
+					}	
+				}
 			}
 			FileOutputStream outputStream = new FileOutputStream(new File(reportpath));
 			FileOutputStream evidenceStream = new FileOutputStream(new File(evidencepath));
@@ -140,6 +162,5 @@ public class MyListener implements ISuiteListener {
         row.createCell(4).setCellValue(status);
         row.createCell(6).setCellValue("Evidence in Doc File");
 	}
-	
-	
+
 }
